added optimize mode
Ok so redefining these was a nightmare, it kept corrupting the image. I
was worried that this was fucked.

Luckily though there is another trick. make each component only have one
slot, and have that slot be a payload of type class. The class contains
the data so it gets redefined (which it is fine with).

So we don't lose out we also have optimize mode, which uses the same
system as before.

This is how you switch to optimize

```
     (def-component test9-non-opt ()
	 ((x 0s0 :type single-float)
	  (y 0s0 :type single-float))
       (update :x (+ x 1) :y (+ y 2)))

     (def-component (test9-opt :optimize t) ()
	 ((x 0s0 :type single-float)
	  (y 0s0 :type single-float))
       (update :x (+ x 1) :y (+ y 2)))
```