gensym function-names bug
I was getting some weird errors with `with-<component-name>` saying
functions were missing, when they werent. Even odder it was working fine
from the repl but not in the system body.

Changing the names to be regular symbols in the package hasty-hidden
fixes the issue, with the nice side effect that I'm not poluting the
function space with functiosn that can never be reached again on each recompile